requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.0.0
tensorflow==2.13.0
torch==2.0.1
fastapi==0.100.0
uvicorn==0.23.2
pydantic==2.1.1
mlflow==2.7.1
pytest==7.4.0
pytest-cov==4.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pymongo==4.5.0
sqlalchemy==2.0.20
plotly==5.16.1
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
pyyaml==6.0.1
black==23.9.1
flake8==6.0.0
isort==5.12.0
